# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BRAND TOPLIST API Documentation",
        "description": "This is the API documentation for the Brand Toplist application",
        "contact": {
            "email": "yimhelgerman@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get list of brands",
                "operationId": "41ee8c5ca737e71782e6d058a4c4a36f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "brand_name": {
                                                "type": "string"
                                            },
                                            "brand_image": {
                                                "type": "string"
                                            },
                                            "rating": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Create a new brand",
                "operationId": "ee7e7f8fcd9a6991653c61e4570359b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_name",
                                    "brand_image",
                                    "rating"
                                ],
                                "properties": {
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Nike"
                                    },
                                    "brand_image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Brand created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "brand_name": {
                                            "type": "string"
                                        },
                                        "brand_image": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/brands/{brand_id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get a specific brand",
                "operationId": "bad7371f5d0b143b675d79ccb0765d65",
                "parameters": [
                    {
                        "name": "brand_id",
                        "in": "path",
                        "description": "ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Brand found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "brand_name": {
                                            "type": "string"
                                        },
                                        "brand_image": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Brands"
                ],
                "summary": "Update a specific brand",
                "operationId": "384479b7d4252d258426de634c40c14c",
                "parameters": [
                    {
                        "name": "brand_id",
                        "in": "path",
                        "description": "ID of the brand to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Adidas"
                                    },
                                    "brand_image": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Brand updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "brand_name": {
                                            "type": "string"
                                        },
                                        "brand_image": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Delete a specific brand",
                "operationId": "2655c58ba4abdea27bc427bbf5a92d46",
                "parameters": [
                    {
                        "name": "brand_id",
                        "in": "path",
                        "description": "ID of the brand to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Brand deleted successfully"
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Brands",
            "description": "Brands"
        }
    ]
}